Extract Tls12SelectorTest
